Template Inheritence Added

+++ model1/utils.py
@@ -1,10 +1,11 @@
 class DLLNode:
-    def __init__(self, key, val):
+    def __init__(self, key, val, language):
         self.val = val
         self.key = key
         self.prev = None
         self.next = None
         self.freq=1
+        self.language=language
 
 class LRU_Cache:
     def __init__(self, capacity):
@@ -12,8 +13,8 @@ class LRU_Cache:
         # Initialize all variable
         self.capacity = capacity
         self.map = {}
-        self.head = DLLNode(0, 0)
-        self.tail = DLLNode(0, 0)
+        self.head = DLLNode(0, 0,"")
+        self.tail = DLLNode(0, 0,"")
         self.head.next = self.tail
         self.tail.prev = self.head
         self.count = 0
@@ -28,26 +29,30 @@ class LRU_Cache:
         node.prev = self.head
         self.head.next = node
     
-    def get(self, key):
+    def get(self, key,language):
         if key in self.map:
             node = self.map[key]
-            result = node.val
-            self.deleteNode(node)
-            self.addToHead(node)
-            print('Got the value : {} for the key: {}'.format(result, key, node.freq))
-            return result
+            if node.language==language:
+                result = node.val
+                self.deleteNode(node)
+                self.addToHead(node)
+                print('Got the value : {} for the key: {}'.format(result, key, node.freq))
+                return result
+            else:
+                return -1
         print('Did not get any value for the key: {}'.format(key))
         return -1
     
-    def set(self, key, value):
+    def set(self, key, value, language):
         print('going to set the (key, value) : ( {}, {})'.format(key, value))
         if key in self.map:
             node = self.map[key]
-            node.val = value
-            self.deleteNode(node)
-            self.addToHead(node)
+            if node.language==language:
+                node.val = value
+                self.deleteNode(node)
+                self.addToHead(node)
         else:
-            node = DLLNode(key, value)
+            node = DLLNode(key, value, language)
             self.map[key] = node
             if self.count < self.capacity:
                 self.count += 1
@@ -65,8 +70,8 @@ class LFU_Cache:
         self.map = {}  # Stores key-node mapping
         self.freq_map = {}  # Stores frequency-doubly linked list mapping
         self.min_freq = 0  # Track the minimum frequency
-        self.head = DLLNode(0, 0)  # Doubly linked list head
-        self.tail = DLLNode(0, 0)  # Doubly linked list tail
+        self.head = DLLNode(0, 0,"")  # Doubly linked list head
+        self.tail = DLLNode(0, 0,"")  # Doubly linked list tail
         self.head.next = self.tail
         self.tail.prev = self.head
 
@@ -87,28 +92,32 @@ class LFU_Cache:
         node.freq += 1
         self.deleteNode(node)
         if node.freq not in self.freq_map:
-            self.freq_map[node.freq] = DLLNode(0, 0)
+            self.freq_map[node.freq] = DLLNode(0, 0, "")
         self.addToHead(node)
 
-    def get(self, key):
+    def get(self, key,language):
         if key in self.map:
             node = self.map[key]
-            result = node.val
-            self.increaseFrequency(node)
-            print('Got the value: {} for the key: {} and frequency: {}'.format(result, key, node.freq))
-            return result
+            if node.language==language:
+                result = node.val
+                self.increaseFrequency(node)
+                print('Got the value: {} for the key: {} and frequency: {}'.format(result, key, node.freq))
+                return result
+            else:
+                return -1
         print('Did not get any value for the key: {}'.format(key))
         return -1
 
-    def set(self, key, value):
+    def set(self, key, value, language):
         print('Going to set the (key, value): ({}, {})'.format(key, value))
         if key in self.map:
             node = self.map[key]
-            node.val = value
-            self.increaseFrequency(node)
+            if node.language==language:
+                node.val = value
+                self.increaseFrequency(node)
         else:
             # Create a new node
-            node = DLLNode(key, value)
+            node = DLLNode(key, value,language)
             self.map[key] = node
             if len(self.map) > self.capacity:
                 # Remove the least frequently used item
@@ -119,5 +128,5 @@ class LFU_Cache:
                 self.deleteNode(to_remove)
             self.min_freq = 1
             if self.min_freq not in self.freq_map:
-                self.freq_map[self.min_freq] = DLLNode(0, 0)
+                self.freq_map[self.min_freq] = DLLNode(0, 0,"")
             self.addToHead(node)
